Add axis parameter to dataframe.diff (#774)
Since pandas dataframe support axis parameter in dataframe.diff.

<img width="772" alt="스크린샷 2019-09-12 오후 1 47 32" src="https://user-images.githubusercontent.com/44108233/64754921-30949300-d564-11e9-8ae6-8f7e52f18951.png">

I think we need to support it either. (And it was already commented as TODO)

In addition, now the nunique function only support the axis parameter as 0 (without 'index'), also fix it.

<img width="922" alt="스크린샷 2019-09-12 오후 1 53 41" src="https://user-images.githubusercontent.com/44108233/64755104-fb3c7500-d564-11e9-83cb-93226ab8b4b7.png">


And currently we only support axis 0, or axis 1 for some functions, so i plan to implement it soon.