add conda-forge channel (#1288)
By default, Anaconda does not have `conda-forge` channel in the config, causing

```
conda install koalas
```

from [Getting started](https://koalas.readthedocs.io/en/latest/getting_started/install.html) to fail. The issue is not apparent for the people not accustomed to `conda` as demonstrated in [this SO post](https://stackoverflow.com/questions/60263669/how-to-install-koalas-with-conda/60263951#60263951).