read_csv: Implement reading of number of rows (#1656)
Implement reading of number of rows (nrows) in read_csv by using spark's limit.

On the first read, this does not seem to help much with regards to reading speed, because `inferSchema` is True and spark seems to scan over the full data anyway (see e.g. [here](https://stackoverflow.com/questions/44277019/how-to-read-only-n-rows-of-large-csv-file-on-hdfs-using-spark-csv-package)). It would help to use it as a similar parameter to `read_csv` than in pandas to make the API more compatible.