Set the lowerbound of numpydoc (#1866)
This PR proposes to set the lower bound of numpydoc. The main reason to set the bound to 0.8.0 was that `self` argument appeared in the API references (see https://github.com/databricks/koalas/issues/439). This seems fixed now from 1.1.0+

There are a couple of methods not being found in "See Also" via numpydoc but I think they are trivial issues.